Int unification (#12)
* service id int -> int64

* rest int -> int64, int32 -> int64

* int64 convertors

* pg migrations int -> big int

+++ internal/api/category/create.go
@@ -17,7 +17,7 @@ func (categoryImpl *Implementation) Create(ctx context.Context, req *desc.Create
 	}
 
 	res := &desc.CreateResponse{
-		Id: int64(id),
+		Id: id,
 	}
 
 	return res, nil
