+++ internal/api/converter/subcategory.go
@@ -9,9 +9,9 @@ import (
 
 func SubcategoryToGRPC(subcategory *model.Subcategory) *desc.GetResponse {
 	res := &desc.GetResponse{
-		Id:         int64(subcategory.ID),
+		Id:         subcategory.ID,
 		Name:       subcategory.Name,
-		CategoryId: int64(subcategory.CategoryID),
+		CategoryId: subcategory.CategoryID,
 		CreatedAt:  timestamppb.New(subcategory.CreatedAt),
 		UpdatedAt:  nil,
 	}
@@ -24,7 +24,7 @@ func SubcategoryToGRPC(subcategory *model.Subcategory) *desc.GetResponse {
 }
 
 func NewSubcategoryFromGRPC(req *desc.CreateRequest) *model.NewSubcategory {
-	newSubcategory := &model.NewSubcategory{Name: req.Name, CategoryID: int(req.CategoryId)}
+	newSubcategory := &model.NewSubcategory{Name: req.Name, CategoryID: req.CategoryId}
 
 	return newSubcategory
 }
@@ -38,7 +38,7 @@ func UpdatedSubcategoryFromGRPC(req *desc.UpdateRequest) *model.UpdatedSubcatego
 	}
 
 	if req.CategoryId != nil {
-		domainId := int(req.CategoryId.GetValue())
+		domainId := req.CategoryId.GetValue()
 		updatedSubcategory.CategoryID = &domainId
 	}
