+++ internal/api/question/create.go
@@ -15,7 +15,7 @@ func (questionImpl *Implementation) Create(ctx context.Context, req *desc.Create
 	}
 
 	res := &desc.CreateResponse{
-		Id: int64(id),
+		Id: id,
 	}
 
 	return res, nil
