+++ internal/api/questionset/create.go
@@ -15,7 +15,7 @@ func (questionSetImpl *Implementation) Create(ctx context.Context, req *desc.Cre
 	}
 
 	res := &desc.CreateResponse{
-		Id: int64(id),
+		Id: id,
 	}
 
 	return res, nil
