+++ internal/service/category/get.go
@@ -6,7 +6,7 @@ import (
 	"github.com/Kosfedev/learn_go/internal/model"
 )
 
-func (qcs *serv) Get(ctx context.Context, id int) (*model.Category, error) {
+func (qcs *serv) Get(ctx context.Context, id int64) (*model.Category, error) {
 	category, err := qcs.repo.Get(ctx, id)
 	if err != nil {
 		return nil, err
