+++ internal/service/domain/delete.go
@@ -4,7 +4,7 @@ import (
 	"context"
 )
 
-func (qds *serv) Delete(ctx context.Context, id int) error {
+func (qds *serv) Delete(ctx context.Context, id int64) error {
 	err := qds.repo.Delete(ctx, id)
 	if err != nil {
 		return err
