+++ internal/service/domain/update.go
@@ -6,7 +6,7 @@ import (
 	"github.com/Kosfedev/learn_go/internal/model"
 )
 
-func (qds *serv) Update(ctx context.Context, id int, updatedDomain *model.UpdatedDomain) error {
+func (qds *serv) Update(ctx context.Context, id int64, updatedDomain *model.UpdatedDomain) error {
 	err := qds.repo.Update(ctx, id, updatedDomain)
 	if err != nil {
 		return err
