+++ internal/service/question/delete.go
@@ -4,7 +4,7 @@ import (
 	"context"
 )
 
-func (qs *serv) Delete(ctx context.Context, id int) error {
+func (qs *serv) Delete(ctx context.Context, id int64) error {
 	err := qs.questionRepo.Delete(ctx, id)
 	if err != nil {
 		return err
