+++ internal/service/questionset/delete.go
@@ -2,7 +2,7 @@ package questionset
 
 import "context"
 
-func (qss *serv) Delete(ctx context.Context, id int) error {
+func (qss *serv) Delete(ctx context.Context, id int64) error {
 	err := qss.repo.Delete(ctx, id)
 	if err != nil {
 		return err
